Fix bug in link navigation code Fix bug which caused crash on startup on F100 Cleaned up button code Hide cursor sooner
git-svn-id: https://gmenu2x.svn.sourceforge.net/svnroot/gmenu2x@128 bad23c4b-591a-0410-b480-ec8053e4ca07

+++ src/menu.cpp
@@ -314,7 +314,7 @@ void Menu::linkRight() {
 void Menu::linkUp() {
 	int l = iLink-numCols;
 	if (l<0) {
-		int rows = ceil(sectionLinks()->size()/(double)numCols);
+		uint rows = (uint)ceil(sectionLinks()->size()/(double)numCols);
 		l = (rows*numCols)+l;
 		if (l >= (int)sectionLinks()->size())
 			l -= numCols;
@@ -325,8 +325,8 @@ void Menu::linkUp() {
 void Menu::linkDown() {
 	uint l = iLink+numCols;
 	if (l >= sectionLinks()->size()) {
-		uint rows = ceil(sectionLinks()->size()/(double)numCols);
-		uint curCol = ceil((iLink+1)/(double)numCols);
+		uint rows = (uint)ceil(sectionLinks()->size()/(double)numCols);
+		uint curCol = (uint)ceil((iLink+1)/(double)numCols);
 		if (rows > curCol)
 			l = sectionLinks()->size()-1;
 		else
